src/sonde_read/requirements.txt
ament_copyright==0.12.15
ament_flake8==0.12.15
ament_pep257==0.12.15
pyserial==3.5
pytest==6.2.5
rclpy==3.3.20
setuptools==59.6.0
std_msgs==4.9.1
